+++ gems/aws-partitions/lib/aws-partitions.rb
@@ -301,6 +301,7 @@ module Aws
             'ConnectContactLens' => 'contact-lens',
             'ConnectParticipant' => 'participant.connect',
             'ConnectWisdomService' => 'wisdom',
+            'ControlTower' => 'controltower',
             'CostExplorer' => 'ce',
             'CostandUsageReportService' => 'cur',
             'CustomerProfiles' => 'profile',