+++ gems/aws-sdk-core/lib/aws-sdk-ssooidc.rb
@@ -50,6 +50,6 @@ require_relative 'aws-sdk-ssooidc/customizations'
 # @!group service
 module Aws::SSOOIDC
 
-  GEM_VERSION = '3.137.0'
+  GEM_VERSION = '3.138.0'
 
 end