+++ gems/aws-sdk-core/lib/aws-sdk-sts.rb
@@ -50,6 +50,6 @@ require_relative 'aws-sdk-sts/customizations'
 # @!group service
 module Aws::STS
 
-  GEM_VERSION = '3.138.0'
+  GEM_VERSION = '3.139.0'
 
 end