+++ gems/aws-sdk-route53/lib/aws-sdk-route53/types.rb
@@ -155,12 +155,11 @@ module Aws::Route53
     # When creating resource record sets for a private hosted zone, note the
     # following:
     #
-    # * Creating geolocation alias resource record sets or latency alias
-    #   resource record sets in a private hosted zone is unsupported.
-    #
     # * For information about creating failover resource record sets in a
     #   private hosted zone, see [Configuring Failover in a Private Hosted
     #   Zone][1].
+    #
+    # ^
     #
     #
     #
@@ -7388,9 +7387,6 @@ module Aws::Route53
     #
     #   * You can't create an alias resource record set in a private hosted
     #     zone to route traffic to a CloudFront distribution.
-    #
-    #   * Creating geolocation alias resource record sets or latency alias
-    #     resource record sets in a private hosted zone is unsupported.
     #
     #   * For information about creating failover resource record sets in a
     #     private hosted zone, see [Configuring Failover in a Private Hosted